requirements.txt
pyusb~=1.1.1
multipledispatch~=0.6.0
